+++ frontend/src/components/Sequence/Sequence.jsx
@@ -33,9 +33,13 @@ export default class Sequence extends Component {
         className={`card-global ${
           (this.props.isEdit || this.props.isCreating) && this.props.id === this.props.stateId
             ? "card-error"
+            : this.props.id === this.props.currentSeq
+            ? "card-current"
             : "card-default"
         } ${this.props.isDarkMode ? "dark-mode" : ""}`}
-        border={(this.props.isEdit || this.props.isCreating) && this.props.id === this.props.stateId ? "none" : "dark"}
+        border={
+          (this.props.isEdit || this.props.isCreating) && this.props.id === this.props.currentSeq ? "none" : "dark"
+        }
       >
         <Card.Header className={`border-bottom-0 header-custom ${this.props.isDarkMode ? "dark-mode" : ""}`}>
           <Row>